% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{bige}
\alias{bige}
\title{select big amount entries}
\usage{
bige(threshold = 1000, x)
}
\arguments{
\item{threshold}{only adjustments absolutely bigger than threshold}

\item{x}{output of std or equivalent}
}
\description{
select big amount entries
}
