% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{blocksprocess}
\alias{blocksprocess}
\title{process files, extract blocks, expand input blocks as needed
returns blocks ready for ledger}
\usage{
blocksprocess(ff, verbose = FALSE)
}
\arguments{
\item{ff}{character vector of files to process}

\item{verbose}{turn on messages (default = FALSE)}
}
\description{
process files, extract blocks, expand input blocks as needed
returns blocks ready for ledger
}
