% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{blocktest}
\alias{blocktest}
\title{not yet functional test to make sure things are working as expected}
\usage{
blocktest(fil = "passbook.example", n = 0)
}
\arguments{
\item{fil}{alternatively, explicitly give the name of the file to test}

\item{n}{which csvfile to test}
}
\description{
not yet functional test to make sure things are working as expected
}
