% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{ca}
\alias{ca}
\title{see adjustment entries : absolutely bigger than threshold}
\usage{
ca(threshold = 0)
}
\arguments{
\item{threshold}{only adjustments absolutely bigger than threshold}
}
\description{
see adjustment entries : absolutely bigger than threshold
}
