% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{cax}
\alias{cax}
\title{see xpense entries : absolutely bigger than threshold (default:1000)}
\usage{
cax(threshold = 1000)
}
\arguments{
\item{threshold}{only expense adjustments bigger than threshold}
}
\description{
see xpense entries : absolutely bigger than threshold (default:1000)
}
