% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{cb}
\alias{cb}
\title{see both cash and investment entries : not 'bad' accounts and not interesting accounts}
\usage{
cb()
}
\description{
see both cash and investment entries : not 'bad' accounts and not interesting accounts
}
