% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{cby}
\alias{cby}
\title{std utility to see various views of accounts}
\usage{
cby(byf, X = cb())
}
\arguments{
\item{byf}{encoding of by fields}

\item{X}{transactions object}
}
\description{
std utility to see various views of accounts
}
