% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{cf}
\alias{cf}
\title{see fixed deposit instruments : absolutely bigger than threshold (default:1000)}
\usage{
cf()
}
\description{
see fixed deposit instruments : absolutely bigger than threshold (default:1000)
}
