% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{costattach}
\alias{costattach}
\title{assign cost to each transaction}
\usage{
costattach(a, p)
}
\arguments{
\item{a}{transactions object}

\item{p}{prices object}
}
\description{
assign cost to each transaction
}
