% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{csvprocess}
\alias{csvprocess}
\title{process inputs from csv files
  if badcsvfile != '', write malformed lines into badcsvfile}
\usage{
csvprocess(csvfiles, badcsvfile = "")
}
\arguments{
\item{csvfiles}{character vector of csv files to process}

\item{badcsvfile}{name of output csvfile to write malformed lines into}
}
\description{
process inputs from csv files
  if badcsvfile != '', write malformed lines into badcsvfile
}
