% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{ct}
\alias{ct}
\title{see tariff entries :  (Invest, Income, Interest, Tax, Trips, Expenses, Equity)}
\usage{
ct()
}
\description{
see tariff entries :  (Invest, Income, Interest, Tax, Trips, Expenses, Equity)
}
