% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{cx}
\alias{cx}
\title{see xpense entries : absolutely bigger than threshold (default:1000)}
\usage{
cx(threshold = 1000)
}
\arguments{
\item{threshold}{only expense adjustments bigger than threshold}
}
\description{
see xpense entries : absolutely bigger than threshold (default:1000)
}
