% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\docType{data}
\name{iacc}
\alias{iacc}
\title{special accounts for selection}
\format{An object of class \code{character} of length 1.}
\usage{
iacc
}
\description{
special accounts for selection
}
\keyword{datasets}
