% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc-internal.R
\name{inhome}
\alias{inhome}
\title{find input files}
\usage{
inhome(...)
}
\arguments{
\item{...}{strings to append to getwd() to determine location of inputs}
}
\value{
location of input files
}
\description{
find input files
}
