% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{locke}
\alias{locke}
\title{see fixed deposit instruments still locked up: absolutely bigger than threshold (default:1000)}
\usage{
locke(x)
}
\arguments{
\item{x}{output of std or equivalent}
}
\description{
see fixed deposit instruments still locked up: absolutely bigger than threshold (default:1000)
}
