% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc-internal.R
\name{lstr}
\alias{lstr}
\title{for debugging, limit str output to six elements}
\usage{
lstr(...)
}
\arguments{
\item{...}{pass through to str}
}
\description{
for debugging, limit str output to six elements
}
