% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{notacc}
\alias{notacc}
\title{account deselection}
\usage{
notacc(aca, a)
}
\arguments{
\item{aca}{transactions object}

\item{a}{regular expression of accounts to deselect}
}
\description{
account deselection
}
