% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{notefield}
\alias{notefield}
\title{utility to extract note fields}
\usage{
notefield(x, f)
}
\arguments{
\item{x}{object from which to extract notefield}

\item{f}{name of notefield, should appear in the form f:<finfo> in the note}
}
\description{
utility to extract note fields
}
