% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{now}
\alias{now}
\title{pick out transactions around today}
\usage{
now(aca, before = 0, after = 0)
}
\arguments{
\item{aca}{transactions object}

\item{before}{number of days before today to select}

\item{after}{number of days after today to select}
}
\description{
pick out transactions around today
}
