% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc-internal.R
\name{outhome}
\alias{outhome}
\title{find location of output files}
\usage{
outhome(...)
}
\arguments{
\item{...}{relative path for output}
}
\value{
location where output files should be placed
}
\description{
find location of output files
}
