% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{p}
\alias{p}
\title{plot xpenses by category vs time}
\usage{
p(X = cby(y), y = "cval", bigfont = 8)
}
\arguments{
\item{X}{transactions of interest}

\item{y}{value to plot}

\item{bigfont}{font size}
}
\description{
plot xpenses by category vs time
}
