% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{pby}
\alias{pby}
\title{smart plot of extracted data, pie charts, multi-panel plots}
\usage{
pby(...)
}
\arguments{
\item{...}{arguments are passed to cby}
}
\description{
smart plot of extracted data, pie charts, multi-panel plots
}
