% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{posprocess}
\alias{posprocess}
\title{process csv file(s) for positions and prices}
\usage{
posprocess(csvfiles)
}
\arguments{
\item{csvfiles}{character vector of csv files to process}
}
\description{
process csv file(s) for positions and prices
}
