% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{pycsv}
\alias{pycsv}
\title{for debugging and for internal use return all transactions as csv lines}
\usage{
pycsv()
}
\value{
transactions as csv lines
}
\description{
for debugging and for internal use return all transactions as csv lines
}
