% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{pypq}
\alias{pypq}
\title{return all price lines}
\usage{
pypq(pypfile)
}
\arguments{
\item{pypfile}{file containing all price quotes  pyquotes file}
}
\value{
price lines as strings
}
\description{
return all price lines
}
