% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc-internal.R
\name{qw}
\alias{qw}
\title{for internal and for public use, mimic perl qw function}
\usage{
qw(x, split = ",")
}
\arguments{
\item{x}{string}

\item{split}{delimiter to split string x}
}
\description{
for internal and for public use, mimic perl qw function
}
