% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{read.hsbc}
\alias{read.hsbc}
\title{process a csv file downloaded from hsbc}
\usage{
read.hsbc(csvf)
}
\arguments{
\item{csvf}{csv file from hsbc to process}
}
\description{
process a csv file downloaded from hsbc
}
