% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{read.jpm}
\alias{read.jpm}
\title{process a csv file downloaded from jpm}
\usage{
read.jpm(csvf)
}
\arguments{
\item{csvf}{csv file from pm to process
Transfers',}
}
\description{
process a csv file downloaded from jpm
}
