% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{since}
\alias{since}
\title{select transactions on or after date b}
\usage{
since(aca, b = lubridate::floor_date(today - 10, "year"))
}
\arguments{
\item{aca}{transactions object}

\item{b}{date as a string or Date object, defaults to beginning of current year}
}
\description{
select transactions on or after date b
}
