% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{sumt}
\alias{sumt}
\title{general aggregation function, should be applied directly to raw data}
\usage{
sumt(a, byfields = c("account", "units"))
}
\arguments{
\item{a}{transactions object}

\item{byfields}{grouping variables for aggregation}
}
\description{
general aggregation function, should be applied directly to raw data
}
