% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{supd}
\alias{supd}
\title{invoke upd with messages suppressed, update accounting info by reprocessing all input files}
\usage{
supd(reprocess = FALSE)
}
\arguments{
\item{reprocess}{reprocess files even if nothing changes}
}
\description{
invoke upd with messages suppressed, update accounting info by reprocessing all input files
}
