% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{tacctype}
\alias{tacctype}
\title{acctype selection}
\usage{
tacctype(aca, u)
}
\arguments{
\item{aca}{transactions object}

\item{u}{regular expression of acctype to select}
}
\description{
acctype selection
}
