% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{transprocess}
\alias{transprocess}
\title{process csv file(s) for transactions}
\usage{
transprocess(csvfiles)
}
\arguments{
\item{csvfiles}{character vector of csv files to process}
}
\description{
process csv file(s) for transactions
}
