% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{until}
\alias{until}
\title{select transactions on or before date e}
\usage{
until(aca, e = today)
}
\arguments{
\item{aca}{transactions object}

\item{e}{date as a string}
}
\description{
select transactions on or before date e
}
