% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{upd}
\alias{upd}
\title{update accounting info by reprocessing all input files}
\usage{
upd(reprocess = FALSE, verbose = TRUE)
}
\arguments{
\item{reprocess}{reprocess files even if nothing changes}

\item{verbose}{turn on debugging output ?}
}
\description{
update accounting info by reprocessing all input files
}
