% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{valattach}
\alias{valattach}
\title{assign value to each transaction}
\usage{
valattach(a, p)
}
\arguments{
\item{a}{transactions object}

\item{p}{prices object}
}
\description{
assign value to each transaction
}
