% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{xg}
\alias{xg}
\title{compute cumulative statistics}
\usage{
xg(xa, byfields)
}
\arguments{
\item{xa}{input data frame}

\item{byfields}{comma separated string values}
}
\description{
compute cumulative statistics
}
