% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{pby}
\alias{pby}
\title{smart plot of extracted data,
stack plots,
pie charts,
multi-panel plots [deprecated?]}
\usage{
pby(...)
}
\arguments{
\item{...}{cby arguments}
}
\description{
smart plot of extracted data,
stack plots,
pie charts,
multi-panel plots [deprecated?]
}
