% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{pby}
\alias{pby}
\title{plot xpenses by category vs time}
\usage{
pby(X = cby(y), y = "cbyval")
}
\arguments{
\item{X}{transactions of interest}

\item{y}{value to plot}
}
\description{
plot xpenses by category vs time
}
