% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{pyall}
\alias{pyall}
\title{for debugging and for internal use extract all transaction lines}
\usage{
pyall(pos = 0)
}
\arguments{
\item{pos}{which line is of interest (usually for debugging)}
}
\value{
transactions lines as strings invisibly
}
\description{
for debugging and for internal use extract all transaction lines
}
