% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{std}
\alias{std}
\title{'standard loading scheme}
\usage{
std(intrans = pycsv(), inprx = pyprx(), e = today)
}
\arguments{
\item{intrans}{transactions objecct}

\item{inprx}{price info}
}
\description{
'standard loading scheme
}
