% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myacc.R
\name{supd}
\alias{supd}
\title{invoke upd with messages suppressed, update accounting info by reprocessing all input files}
\usage{
supd(acceptalerts = FALSE)
}
\arguments{
\item{acceptalerts}{should badly processed csv lines be written to badcsv files}
}
\description{
invoke upd with messages suppressed, update accounting info by reprocessing all input files
}
